move auxiliary helpers to a support folder
