+++ spec/support/coverage_helper.rb
