+++ spec/support/plugins_helper.rb
